ldap-server/flask-app/requirements.txt
flask==2.1.3
werkzeug==2.1.2
python-ldap==3.4.3
gunicorn==20.1.0